Moved scrollToThread function to separate lib module
