+++ components/common/CharmEditor/components/PageThread.tsx
@@ -11,12 +11,12 @@ import { forwardRef, MouseEvent, useState } from 'react';
 import DeleteIcon from '@mui/icons-material/DeleteOutlined';
 import EditIcon from '@mui/icons-material/Edit';
 import { PageContent } from 'models';
-import { highlightDomElement } from 'lib/dom/highlight';
 import { removeInlineCommentMark } from 'lib/inline-comments/removeInlineCommentMark';
 import { useEditorViewContext } from '@bangle.dev/react';
 import { DateTime } from 'luxon';
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
 import { usePopupState, bindMenu } from 'material-ui-popup-state/hooks';
+import { scrollToThread } from 'lib/inline-comments/client/scrollToThread';
 import InlineCharmEditor from '../InlineCharmEditor';
 import { checkForEmpty } from '../utils';
 
@@ -109,7 +109,7 @@ function AddCommentCharmEditor (
         <Button
           disabled={disabled || isEmpty}
           size='small'
-          onClick={(e) => {
+          onClick={() => {
             onClick(() => {
               if (commentContent) {
                 addComment(threadId, commentContent);
@@ -128,49 +128,6 @@ interface PageThreadProps {
   threadId: string;
   inline?: boolean;
   showFindButton?: boolean;
-}
-
-function scrollToThread (threadId: string) {
-  // Find the inline-comment with the threadId and scroll into view
-  const threadDocument = document.getElementById(`inline-comment.${threadId}`);
-  if (threadDocument) {
-    let parentElement: HTMLElement | null = null;
-    let element: HTMLElement | null = threadDocument;
-    // Check for highest 5 levels of depth
-    for (let i = 0; i < 5; i++) {
-      element = threadDocument.parentElement;
-      // Get the first paragraph parent element
-      if (element?.classList.contains('bangle-nv-content')) {
-        parentElement = element;
-        break;
-      }
-    }
-
-    requestAnimationFrame(() => {
-      threadDocument.scrollIntoView({
-        behavior: 'smooth'
-      });
-    });
-
-    setTimeout(() => {
-      if (parentElement) {
-        // Need to create a custom element as adding styling to prosemirror-node isn't possible
-        const highlightElement = document.createElement('div');
-        document.body.appendChild(highlightElement);
-        const boundingRect = parentElement.getBoundingClientRect();
-        // Set the location of the custom element
-        highlightElement.style.top = `${boundingRect.top}px`;
-        highlightElement.style.left = `${boundingRect.left}px`;
-        highlightElement.style.width = `${boundingRect.width}px`;
-        highlightElement.style.height = `${boundingRect.height}px`;
-        highlightElement.style.position = 'absolute';
-        highlightDomElement(highlightElement, () => {
-          // Remove the custom element after the highlighting is done
-          document.body.removeChild(highlightElement);
-        });
-      }
-    }, 500);
-  }
 }
 
 const PageThread = forwardRef<HTMLDivElement, PageThreadProps>(({ showFindButton = false, threadId, inline = false }, ref) => {
