+++ components/common/CharmEditor/components/inlineComment/inlineComment.plugins.ts
@@ -1,6 +1,6 @@
 import { Plugin, RawPlugins } from '@bangle.dev/core';
 import { TextSelection } from '@bangle.dev/pm';
-import { highlightDomElement } from 'lib/dom/highlight';
+import { highlightDomElement } from 'lib/dom/client/highlight';
 import { renderSuggestionsTooltip, SuggestTooltipPluginKey } from '../@bangle.dev/tooltip/suggest-tooltip';
 
 export function plugin (): RawPlugins {
