+++ lib/dom/client/highlight.ts
